requirements.txt
discord.py-self
aiosqlite
python-dotenv
